Restructured and Modularized the Navbar into several components
