+++ .gitignore
@@ -17,6 +17,8 @@
 .env.development.local
 .env.test.local
 .env.production.local
+.github
+.vscode
 
 npm-debug.log*
 yarn-debug.log*
