+++ src/App.js
@@ -4,7 +4,7 @@ import React, { useState } from "react";
 import About from "./Component/About";
 import Footer from "./Component/Footer";
 import Home from "./Component/Home";
-import Navbar from "./Component/Navbar";
+import Navbar from "./Component/Navbar/Navbar";
 import BookMark from "./Component/BookMark";
 import NotFound from "./Component/NotFound"
 import BackToTopButton from "./Component/BackToTopButton";
@@ -14,7 +14,6 @@ function App() {
   return (
     <>
       <Navbar setSearchQuery={setSearchQuery} />
-
       <Routes>
         <Route path="/" element={<Home searchQuery={searchQuery} />}></Route>
         <Route path="/bookmark" element={<BookMark />}></Route>
